Update pmem2bench: use TLS for read/write buffers

+++ benchmarks/pmem2bench.sh
@@ -1,11 +1,13 @@
 #!/usr/bin/bash
 
 OUTPUT_DIR_PATH="$HOME/bench_results/pmem2bench"
-PMEM2BENCH_EXE="$HOME/work/spack-playground/build-RelWithDebInfo/bin/pmem2bench"
-POOL_PATH=/mnt/pmem0/$USER/pmem2bench.pool
+PMEM2BENCH_EXE="sudo $HOME/work/spack-playground/build-RelWithDebInfo/bin/pmem2bench"
+SOURCE_TYPE="devdax"
+# POOL_PATH=/mnt/pmem0/$USER/pmem2bench.pool
+POOL_PATH=/dev/dax0.0
 POOL_SIZE=8G
 TOTAL_SIZE=8G
-LABEL=$(date --iso-8601=seconds)-8G-all
+LABEL=$(date --iso-8601=seconds)-8G-devdax-tls_buf
 
 access_pattern_list=(
   ""
@@ -29,6 +31,7 @@ mkdir -p "$OUTPUT_DIR_PATH"
 for store_type in "${store_type_list[@]}"; do
 for access_pattern in "${access_pattern_list[@]}"; do
 for nthreads in 1 2 4 8 16; do
+# for nthreads in 16; do
 for block_size in 64 128 256 512 1K 2K 4K 8K 16K 32K 64K 128K 256K 512K 1M; do
 # block_size_byte=$(echo $block_size | numfmt --from=iec)
 # stripe_size=$((block_size_byte * nthreads))
@@ -40,7 +43,8 @@ echo $OUTPUT_FILE
 
   # --stripe $stripe_size
 $PMEM2BENCH_EXE \
-  --file $POOL_PATH \
+  --source $SOURCE_TYPE \
+  --path $POOL_PATH \
   --file_size $POOL_SIZE \
   --total $TOTAL_SIZE \
   --stripe $TOTAL_SIZE \
@@ -55,7 +59,9 @@ $PMEM2BENCH_EXE \
 done
 done
 
+if [ $SOURCE_TYPE = "fsdax" ] ; then
 rm $POOL_PATH
+fi
 
 done
 done
